apollo_state all but one test and I'm beat
need to mock out the localStorage object
