+++ package.json
@@ -51,6 +51,7 @@
     "jquery": "^3.4.1",
     "localStorage": "^1.0.4",
     "mini-css-extract-plugin": "0.9.0",
+    "mock-apollo-client": "^0.4.0",
     "node-sass": "^4.9.4",
     "optimize-css-assets-webpack-plugin": "5.0.3",
     "pnp-webpack-plugin": "1.6.0",
