+++ src/App.spec.js
@@ -1,20 +1,20 @@
-import Cookies from 'universal-cookie';
-import cookieObject from './cookieObject';
+import localStorage from 'localStorage';
 
-const cookies = new Cookies();
-const cookieExists = cookies.get('groupy')
+const ls = JSON.parse(localStorage.getItem('groupy'));
+const userID = ls.userID;
+
 const App = require('./App')
 
 test('proper profile data is correct', () => {
-  expect(cookieObject).toMatchSnapshot();
-  expect(cookieObject.userID).toEqual('5e62ade15014b544a50af943');
+  expect(ls).toMatchSnapshot();
+  expect(userID).toEqual('5e62ade15014b544a50af943');
 });
 
-test('new cookies', () => {
-  expect(cookies.HAS_DOCUMENT_COOKIE).toBe(true);
-});
+// test('new cookies', () => {
+//   expect(cookies.HAS_DOCUMENT_COOKIE).toBe(true);
+// });
 
-test('cookieObject should have properties token and userID', () => {
-  expect(cookieObject).toHaveProperty('token');
-  expect(cookieObject).toHaveProperty('userID');
-});
+// test('cookieObject should have properties token and userID', () => {
+//   expect(cookieObject).toHaveProperty('token');
+//   expect(cookieObject).toHaveProperty('userID');
+// });
